+++ game/square.py
@@ -3,21 +3,27 @@ from game.tile import Tile
 
 class Square:
     def __init__(self):
-        self.multiplier = 1
+        self.letter_multiplier = 1
+        self.word_multiplier = None
         self.letter = None
         self.multiplier_type = None
+        self.multipler_up = True
 
     def has_letter(self):
         return self.letter is not None
 
     def has_multiplier(self):
-        return self.multiplier > 1
+        return self.letter_multiplier > 1
 
     def get_multiplier_type(self):
         return self.multiplier_type
 
     def get_multiplier(self):
-        return self.multiplier
+        return self.letter_multiplier
+
+    def multiplier_is_up(self):
+        if self.has_letter():
+            self.multipler_up = False
 
     def get_letter(self):
         return self.letter
@@ -25,13 +31,20 @@ class Square:
     def put_letter(self, letter):
         if not self.has_letter():
             self.letter = letter
+        self.multiplier_is_up()
+
+    def individual_score(self):
+        return self.letter.get_value() * self.letter_multiplier
 
     def set_multiplier_type(self, word):
         self.multiplier_type = word
 
     def set_multiplier(self, amount):
-        self.multiplier = amount
+        self.letter_multiplier = amount
 
     def set_letter(self, tile):
         self.letter = tile
+
+    def set_word_multiplier(self, amount):
+        self.word_multiplier = amount